requirements.txt
```text
Pillow>=10.0.0